Improve UI.

+++ src/components/Label.tsx
@@ -11,17 +11,19 @@ function Label({ label }: { label: LabelType }) {
         refetchQueries: [{ query: GET_LABELS }],
     })
     function handleClick() {
-        deleteTask({ variables: { id } })
+        if (confirm(`Delete label ${label.name} ?`)) {
+            deleteTask({ variables: { id } })
+        }
     }
     if (loading) return <p>Submitting...</p>
     if (error) return <p className="text-red-500">Error: {error.message}</p>
 
     return (
-        <li className={`my-4 flex justify-evenly`}>
+        <li className={`my-2 flex justify-evenly bg-stone-200 rounded`}>
             <UpdateLabel label={label} />
             <button
                 onClick={handleClick}
-                className="my-2 py-2 px-2 border-solid border-2 text-blue-500 border-blue-500 hover:border-blue-700 rounded"
+                className="my-2 py-2 px-2 border-solid border-2 text-red-500 border-red-500 hover:border-red-700 rounded"
             >
                 X
             </button>
